code_checker/requirements.txt
Flask==2.3.3
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
easyocr==1.7.1
pytesseract==0.3.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
authlib==1.2.1
requests==2.31.0
Werkzeug==2.3.7
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3